+++ src/bin_evm_proxy/lib_dev/rpc_encodings.ml
@@ -505,6 +505,38 @@ module Get_transaction_by_block_number_and_index = MethodMaker (struct
   let method_ = "eth_getTransactionByBlockNumberAndIndex"
 end)
 
+module Get_uncle_by_block_hash_and_index = MethodMaker (struct
+  open Ethereum_types
+
+  type method_
+
+  type input = block_hash * quantity
+
+  type output = block option
+
+  let input_encoding = Data_encoding.tup2 block_hash_encoding quantity_encoding
+
+  let output_encoding = Data_encoding.option block_encoding
+
+  let method_ = "eth_getUncleByBlockHashAndIndex"
+end)
+
+module Get_uncle_by_block_number_and_index = MethodMaker (struct
+  open Ethereum_types
+
+  type method_
+
+  type input = block_param * quantity
+
+  type output = block option
+
+  let input_encoding = Data_encoding.tup2 block_param_encoding quantity_encoding
+
+  let output_encoding = Data_encoding.option block_encoding
+
+  let method_ = "eth_getUncleByBlockNumberAndIndex"
+end)
+
 module Send_raw_transaction = MethodMaker (struct
   open Ethereum_types
 
@@ -637,6 +669,8 @@ let methods : (module METHOD) list =
     (module Get_transaction_by_hash);
     (module Get_transaction_by_block_hash_and_index);
     (module Get_transaction_by_block_number_and_index);
+    (module Get_uncle_by_block_hash_and_index);
+    (module Get_uncle_by_block_number_and_index);
     (module Send_transaction);
     (module Send_raw_transaction);
     (module Eth_call);