+++ src/kernel_evm/CHANGES.md
@@ -33,6 +33,7 @@
 - Add support for `eth_getBlockTransactionCountByHash` and `eth_getBlockTransactionCountByNumber`. (!10101)
 - Add support for `web3_sha3`. (!10112)
 - Add support for `eth_getUncleCountByBlockHash` and `eth_getUncleCountByBlockNumber`. (!10131)
+- Add support for `eth_getUncleByBlockHashAndIndex` and `eth_getUncleByBlockNumberAndIndex`. (!10134)
 
 
 ### Bug fixes